requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
fastapi==0.104.1           # Web framework
uvicorn[standard]==0.24.0  # ASGI server
sqlalchemy==2.0.23         # Database ORM
alembic==1.12.1           # Database migrations
python-multipart==0.0.6   # File uploads
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4     # Password hashing
python-decouple==3.8       # Environment variables
psycopg2-binary==2.9.9     # PostgreSQL adapter
pillow==10.1.0             # Image processing
aiofiles==23.2.0           # Async file operations
httpx==0.25.2              # HTTP client
pydantic[email]==2.5.0     # Data validation
fastapi-mail==1.4.1        # Email sending
jinja2==3.1.2              # Email templates
redis==5.0.1               # Caching (optional)
celery==5.3.4              # Background tasks (optional)
requests==2.31.0           # HTTP requests